refactor: centralize normal quantile utility
